requirements.txt
pytest-playwright
fake_useragent